[WIP] Move some functions of the offer to a new view file.
